+++ models/contacts.json
@@ -61,14 +61,20 @@
   },
   {
     "id": "JJt7SbqM8PWRl2fxR4Gpo",
-    "name": "Alex",
-    "email": "alex@mail.com",
-    "phone": "111111"
-  },
-  {
-    "id": "5QsTB_ibxRKZIeijwmMEj",
     "name": "Pet9",
     "email": "example@mail.com",
     "phone": "777888777"
+  },
+  {
+    "id": "SeFAE9dZYTeuAodxeGIEN",
+    "name": "Den",
+    "email": "example@mail.com",
+    "phone": "1"
+  },
+  {
+    "id": "OAsmPqFRzcJRI5IAUaSpT",
+    "name": "Jack",
+    "email": "example@mail.com",
+    "phone": "1110000"
   }
 ]
