+++ routes/api/contacts.js
@@ -2,7 +2,10 @@ const express = require("express");
 
 const router = express.Router();
 
-const Joi = require("joi");
+const {
+  postValidation,
+  putValidation,
+} = require("../../utils/validation/middlewareValidation.js");
 
 const {
   listContacts,
@@ -23,12 +26,12 @@ router.get("/", async (req, res, next) => {
 });
 
 router.get("/:contactId", async (req, res, next) => {
-  const id = req.params.contactId;
+  const { contactId } = req.params;
 
   try {
-    const contactById = await getContactById(id);
+    const contactById = await getContactById(contactId);
     if (contactById) {
-      res.status(200).json({ contactById, message: "success response" });
+      return res.status(200).json({ contactById, message: "success response" });
     }
     res.status(404).json({ message: "Not found" });
   } catch (error) {
@@ -36,67 +39,35 @@ router.get("/:contactId", async (req, res, next) => {
   }
 });
 
-router.post("/", async (req, res, next) => {
+router.post("/", postValidation, async (req, res, next) => {
   try {
-    const schema = Joi.object({
-      name: Joi.string().alphanum().min(3).max(30).required(),
-      email: Joi.string()
-        .email({
-          minDomainSegments: 2,
-          tlds: { allow: ["com", "net"] },
-        })
-        .required(),
-      phone: Joi.number().integer().required(),
-    });
-
-    const validationResult = schema.validate(req.body);
-    if (validationResult.error) {
-      return res
-        .status(404)
-        .json({ message: validationResult.error.details[0].message });
-    }
-
     const addedContact = await addContact(req.body);
     res.status(201).json({ addedContact, message: "contatct added" });
-  } catch (error) {}
+  } catch (error) {
+    console.log(error.message);
+  }
 });
 
 router.delete("/:contactId", async (req, res, next) => {
   try {
-    const id = req.params.contactId;
-    const deletedContact = await removeContact(id);
+    const { contactId } = req.params;
+    const deletedContact = await removeContact(contactId);
     if (deletedContact) {
       return res.status(200).json({ message: "contact deleted" });
     }
     return res.status(404).json({ message: "Not found" });
-  } catch (error) {}
+  } catch (error) {
+    console.log(error.message);
+  }
 });
 
-router.put("/:contactId", async (req, res, next) => {
+router.put("/:contactId", putValidation, async (req, res, next) => {
   try {
-    const id = req.params.contactId;
+    const { contactId } = req.params;
     const body = req.body;
 
-    const schema = Joi.object({
-      name: Joi.string().alphanum().min(3).max(30).required(),
-      email: Joi.string()
-        .email({
-          minDomainSegments: 2,
-          tlds: { allow: ["com", "net"] },
-        })
-        .required(),
-      phone: Joi.number().integer().required(),
-    });
-
-    const validationResult = schema.validate(req.body);
-    if (validationResult.error) {
-      return res
-        .status(404)
-        .json({ message: validationResult.error.details[0].message });
-    }
-
     if (body) {
-      const changedContact = await updateContact(id, body);
+      const changedContact = await updateContact(contactId, body);
       if (changedContact) {
         return res
           .status(200)
@@ -105,7 +76,9 @@ router.put("/:contactId", async (req, res, next) => {
       return res.status(404).json({ message: "Not found" });
     }
     res.status(400).json({ message: "missing fields" });
-  } catch (error) {}
+  } catch (error) {
+    console.log(error.message);
+  }
 });
 
 module.exports = router;
